requirements.txt
# Core Framework
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0

# AI and ML
anthropic==0.25.1
sentence-transformers==2.2.2
chromadb==0.4.22
numpy==1.24.3
scikit-learn==1.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.2
python-magic==0.4.27

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database and Storage
sqlite3
sqlalchemy==2.0.25

# Utilities
requests==2.31.0
python-dateutil==2.8.2
uuid
json5==0.9.14
tiktoken==0.5.2

# Development and Testing
pytest==7.4.4
black==23.12.1
flake8==7.0.0

# Security
werkzeug==3.0.1
